Moved the logic package out of app into new module
