+++ app/build.gradle
@@ -33,6 +33,7 @@ android {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation project(path: ':core')
     def lifecycle_version = "2.3.1"
     def arch_version = "2.1.0"
     def hilt_version = '2.35'
