+++ app/src/main/java/com/trueproof/trueproof/activities/UserAgreementActivity.java
@@ -1,22 +1,12 @@
 package com.trueproof.trueproof.activities;
 
-import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.annotation.SuppressLint;
-import android.content.res.Resources;
 import android.os.Bundle;
-import android.text.Html;
-import android.text.Spanned;
 import android.view.View;
-import android.webkit.WebView;
 import android.widget.Button;
-import android.widget.TextView;
 
 import com.trueproof.trueproof.R;
-import com.trueproof.trueproof.logic.Proofing;
-
-import java.io.InputStream;
 
 import static com.trueproof.trueproof.activities.SignUpActivity.agreementFromUser;
