+++ core/src/main/java/com/trueproof/core/logic/ProofLogic.java
@@ -1,0 +1,4 @@
+package com.trueproof.core.logic;
+
+public class ProofLogic {
+}
