+++ settings.gradle
@@ -1,2 +1,4 @@
 include ':app'
 rootProject.name = "True Proof"
+include ':core'
+project(":core").projectDir = new File("core")
